% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmstBayespara-package.R
\docType{package}
\name{rmstBayespara-package}
\alias{rmstBayespara-package}
\alias{rmstBayespara}
\title{The 'rmstBayespara' package.}
\description{
The parametric Bayes regression models using 'Stan' for restricted mean survival time (RMST).
The package implement the model estimation described in Hanada and Kojima (2024).
The posterior sample, Widely Applicable Information Criterion (WAIC), and
Efficient Leave-One-Out Cross-Validation (loo) for some parametric models of
survival data can be available. The RMST under these parametric models can be
computed from the obtained posterior samples.
}
\references{
Hanada, K., & Kojima, M. (2024). Bayesian Parametric Methods for Deriving Distribution of Restricted Mean Survival Time. arXiv e-prints, arXiv-2406.
}
\author{
\strong{Maintainer}: Keisuke Hanada \email{keisuke.hanada.87@gmail.com} (\href{https://orcid.org/0000-0002-1444-0280}{ORCID})

Authors:
\itemize{
  \item Masahiro Kojima (\href{https://orcid.org/0000-0003-0867-7692}{ORCID})
}

}
